+++ packages/api-client/src/asset/AssetAPI.ts
@@ -18,15 +18,12 @@
  */
 
 import axios, {AxiosRequestConfig} from 'axios';
-import logdown from 'logdown';
 
-import {LogFactory} from '@wireapp/commons';
-import {QualifiedConversationId} from '@wireapp/protocol-messaging';
-
-import {AssetUploadData, PostAssetsResponseSchema} from './AssetAPI.schema';
 import {AssetRetentionPolicy} from './AssetRetentionPolicy';
+import {AssetUploadData} from './AssetUploadData';
 import {isValidToken, isValidUUID} from './AssetUtil';
 
+import {BackendFeatures} from '../APIClient';
 import {
   BackendError,
   handleProgressEvent,
@@ -37,7 +34,7 @@ import {
 } from '../http/';
 import {base64MD5FromBuffer, concatToBuffer} from '../shims/node/buffer';
 import {unsafeAlphanumeric} from '../shims/node/random';
-import {RequestCancellationError} from '../user';
+import {QualifiedId, RequestCancellationError} from '../user';
 
 export interface CipherOptions {
   /** Set a custom algorithm for encryption */
@@ -47,16 +44,16 @@ export interface CipherOptions {
 }
 
 export interface AssetAuditData {
-  conversationId: QualifiedConversationId;
+  convId: QualifiedId;
   filename: string;
   filetype: string;
 }
+
 export interface AssetOptions extends CipherOptions {
   public?: boolean;
   retention?: AssetRetentionPolicy;
   /** If given, will upload an asset that can be shared in a federated env */
   domain?: string;
-  /** Used for identifying the conversation the asset belongs to */
   auditData?: AssetAuditData;
 }
 
@@ -65,12 +62,21 @@ export interface AssetResponse {
   mimeType?: string;
 }
 
-export class AssetAPI {
-  private readonly logger: logdown.Logger;
+const ASSET_URLS = {
+  ASSET_V3_URL: '/assets/v3',
+  ASSET_V4_URL: '/assets/v4',
+  ASSET_SERVICE_URL: '/bot/assets',
+  ASSET_V2_URL: '/otr/assets',
+  ASSET_V2_CONVERSATION_URL: '/conversations',
+  ASSET_V1_URL: '/assets',
+  ASSETS_URL: '/assets',
+} as const;
 
-  constructor(private readonly client: HttpClient) {
-    this.logger = LogFactory.getLogger('@wireapp/api-client/AssetAPI');
-  }
+export class AssetAPI {
+  constructor(
+    private readonly client: HttpClient,
+    private readonly backendFeatures: BackendFeatures,
+  ) {}
 
   private getAssetShared(
     assetUrl: string,
@@ -134,7 +140,7 @@ export class AssetAPI {
       public: boolean;
       retention: AssetRetentionPolicy;
       domain?: string;
-      convId?: QualifiedConversationId;
+      convId?: QualifiedId;
       filename?: string;
       filetype?: string;
     } = {
@@ -144,7 +150,7 @@ export class AssetAPI {
     };
 
     if (options?.auditData) {
-      metadataObject.convId = options.auditData.conversationId;
+      metadataObject.convId = options.auditData.convId;
       metadataObject.filename = options.auditData.filename;
       metadataObject.filetype = options.auditData.filetype;
     }
@@ -182,12 +188,6 @@ export class AssetAPI {
     const handleRequest = async (): Promise<AssetUploadData> => {
       try {
         const response = await this.client.sendRequest<AssetUploadData>(config);
-        const validation = PostAssetsResponseSchema.safeParse(response.data);
-
-        if (!validation.success) {
-          this.logger.warn('Asset upload response validation failed:', validation.error);
-        }
-
         return response.data;
       } catch (error) {
         if ((error as BackendError).message === SyntheticErrorLabel.REQUEST_CANCELLED) {
@@ -203,7 +203,128 @@ export class AssetAPI {
     };
   }
 
-  getAsset(
+  getAssetV1(
+    assetId: string,
+    conversationId: string,
+    forceCaching: boolean = false,
+    progressCallback?: ProgressCallback,
+  ) {
+    if (!isValidUUID(assetId)) {
+      throw new TypeError(`Expected asset ID "${assetId}" to only contain alphanumeric values and dashes.`);
+    }
+    if (!isValidUUID(conversationId)) {
+      throw new TypeError(
+        `Expected conversation ID "${conversationId}" to only contain alphanumeric values and dashes.`,
+      );
+    }
+
+    const cancelSource = axios.CancelToken.source();
+    const config: AxiosRequestConfig = {
+      cancelToken: cancelSource.token,
+      method: 'get',
+      onDownloadProgress: handleProgressEvent(progressCallback),
+      onUploadProgress: handleProgressEvent(progressCallback),
+      params: {
+        conv_id: conversationId,
+      },
+      responseType: 'arraybuffer',
+      url: `${ASSET_URLS.ASSET_V1_URL}/${assetId}`,
+    };
+
+    if (forceCaching) {
+      config.params.forceCaching = forceCaching;
+    }
+
+    const handleRequest = async (): Promise<AssetResponse> => {
+      try {
+        const response = await this.client.sendRequest<ArrayBuffer>(config);
+        return {
+          buffer: response.data,
+          mimeType: response.headers['content-type'],
+        };
+      } catch (error) {
+        if ((error as BackendError).message === SyntheticErrorLabel.REQUEST_CANCELLED) {
+          throw new RequestCancellationError('Asset download got cancelled.');
+        }
+        throw error;
+      }
+    };
+
+    return {
+      cancel: () => cancelSource.cancel(SyntheticErrorLabel.REQUEST_CANCELLED),
+      response: handleRequest(),
+    };
+  }
+
+  getAssetV2(
+    assetId: string,
+    conversationId: string,
+    forceCaching: boolean = false,
+    progressCallback?: ProgressCallback,
+  ) {
+    if (!isValidUUID(assetId)) {
+      throw new TypeError(`Expected asset ID "${assetId}" to only contain alphanumeric values and dashes.`);
+    }
+    if (!isValidUUID(conversationId)) {
+      throw new TypeError(
+        `Expected conversation ID "${conversationId}" to only contain alphanumeric values and dashes.`,
+      );
+    }
+
+    const cancelSource = axios.CancelToken.source();
+    const config: AxiosRequestConfig = {
+      cancelToken: cancelSource.token,
+      method: 'get',
+      onDownloadProgress: handleProgressEvent(progressCallback),
+      onUploadProgress: handleProgressEvent(progressCallback),
+      params: {},
+      responseType: 'arraybuffer',
+      url: `${ASSET_URLS.ASSET_V2_CONVERSATION_URL}/${conversationId}${ASSET_URLS.ASSET_V2_URL}/${assetId}`,
+    };
+
+    if (forceCaching) {
+      config.params.forceCaching = forceCaching;
+    }
+
+    const handleRequest = async (): Promise<AssetResponse> => {
+      try {
+        const response = await this.client.sendRequest<ArrayBuffer>(config);
+        return {
+          buffer: response.data,
+          mimeType: response.headers['content-type'],
+        };
+      } catch (error) {
+        if ((error as BackendError).message === SyntheticErrorLabel.REQUEST_CANCELLED) {
+          throw new RequestCancellationError('Asset download got cancelled.');
+        }
+        throw error;
+      }
+    };
+
+    return {
+      cancel: () => cancelSource.cancel(SyntheticErrorLabel.REQUEST_CANCELLED),
+      response: handleRequest(),
+    };
+  }
+
+  getAssetV3(
+    assetId: string,
+    token?: string | null,
+    forceCaching: boolean = false,
+    progressCallback?: ProgressCallback,
+  ) {
+    const version2 = 2;
+
+    if (!isValidUUID(assetId)) {
+      throw new TypeError(`Expected asset ID "${assetId}" to only contain alphanumeric values and dashes.`);
+    }
+    if (this.backendFeatures.version >= version2) {
+      throw new TypeError('Asset v3 is not supported on backend version 2 or higher.');
+    }
+    return this.getAssetShared(`${ASSET_URLS.ASSET_V3_URL}/${assetId}`, token, forceCaching, progressCallback);
+  }
+
+  getAssetV4(
     assetId: string,
     assetDomain: string,
     token?: string | null,
@@ -220,8 +341,8 @@ export class AssetAPI {
     if (!isValidDomain(assetDomain)) {
       throw new TypeError(`Invalid asset domain ${assetDomain}`);
     }
-
-    return this.getAssetShared(`/assets/${assetDomain}/${assetId}`, token, forceCaching, progressCallback);
+    const assetBaseUrl = this.backendFeatures.version >= 2 ? ASSET_URLS.ASSETS_URL : ASSET_URLS.ASSET_V4_URL;
+    return this.getAssetShared(`${assetBaseUrl}/${assetDomain}/${assetId}`, token, forceCaching, progressCallback);
   }
 
   getServiceAsset(
@@ -234,7 +355,7 @@ export class AssetAPI {
       throw new TypeError(`Expected asset ID "${assetId}" to only contain alphanumeric values and dashes.`);
     }
 
-    const assetBaseUrl = `/bot/assets/${assetId}`;
+    const assetBaseUrl = `${ASSET_URLS.ASSET_SERVICE_URL}/${assetId}`;
     return this.getAssetShared(assetBaseUrl, token, forceCaching, progressCallback);
   }
 
@@ -246,10 +367,12 @@ export class AssetAPI {
    * @param progressCallback? Will be called at every progress of the upload
    */
   postAsset(asset: Uint8Array, options?: AssetOptions, progressCallback?: ProgressCallback) {
-    return this.postAssetShared('/assets', asset, options, progressCallback);
+    const baseUrl = this.backendFeatures.version >= 2 ? ASSET_URLS.ASSETS_URL : ASSET_URLS.ASSET_V3_URL;
+    return this.postAssetShared(baseUrl, asset, options, progressCallback);
   }
 
   postServiceAsset(asset: Uint8Array, options?: AssetOptions, progressCallback?: ProgressCallback) {
-    return this.postAssetShared('/bot/assets', asset, options, progressCallback);
+    const assetBaseUrl = ASSET_URLS.ASSET_SERVICE_URL;
+    return this.postAssetShared(assetBaseUrl, asset, options, progressCallback);
   }
 }