+++ packages/api-client/src/user/User.ts
@@ -35,8 +35,10 @@ export interface User {
   id: string;
   name: string;
   picture?: Picture[];
+  user_groups?: string[];
   qualified_id: QualifiedId;
   service?: ServiceRef;
   team?: string;
+  searchable?: boolean;
   supported_protocols?: ConversationProtocol[];
 }