+++ packages/copy-config/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.8](https://github.com/wireapp/wire-web-packages/compare/@wireapp/copy-config@2.3.7...@wireapp/copy-config@2.3.8) (2025-10-23)
+
+**Note:** Version bump only for package @wireapp/copy-config
+
+## [2.3.7](https://github.com/wireapp/wire-web-packages/compare/@wireapp/copy-config@2.3.6...@wireapp/copy-config@2.3.7) (2025-10-23)
+
+**Note:** Version bump only for package @wireapp/copy-config
+
+## [2.3.6](https://github.com/wireapp/wire-web-packages/compare/@wireapp/copy-config@2.3.5...@wireapp/copy-config@2.3.6) (2025-10-23)
+
+**Note:** Version bump only for package @wireapp/copy-config
+
+## [2.3.5](https://github.com/wireapp/wire-web-packages/compare/@wireapp/copy-config@2.3.4...@wireapp/copy-config@2.3.5) (2025-10-23)
+
+**Note:** Version bump only for package @wireapp/copy-config
+
 ## [2.3.4](https://github.com/wireapp/wire-web-packages/compare/@wireapp/copy-config@2.3.3...@wireapp/copy-config@2.3.4) (2025-09-25)
 
 **Note:** Version bump only for package @wireapp/copy-config